+++ DatabaseStructure/DatabaseStructure.csproj
@@ -51,18 +51,20 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Class1.cs" />
-    <Compile Include="Course.cs" />
-    <Compile Include="Professor.cs" />
+    <Compile Include="AbstractModels\Person.cs" />
+    <Compile Include="DBContext.cs" />
+    <Compile Include="Models\Course.cs" />
+    <Compile Include="Models\Professor.cs" />
+    <Compile Include="Models\School.cs" />
+    <Compile Include="Models\SchoolType.cs" />
+    <Compile Include="Models\Student.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="School.cs" />
-    <Compile Include="SchoolType.cs" />
-    <Compile Include="Student.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
     <None Include="packages.config" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
     <PropertyGroup>
