+++ DatabaseStructure/Models/Course.cs
@@ -1,11 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DatabaseStructure
+namespace DatabaseStructure.Models
 {
     public class Course
     {
@@ -20,6 +19,5 @@ namespace DatabaseStructure
         public virtual Professor CourseProfessor { get; set; }
 
         public virtual ICollection<Student> CourseStudents { get; set; }
-
     }
 }
