+++ DatabaseStructure/Models/SchoolType.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DatabaseStructure.Models
+{
+
+    public class SchoolType
+    {
+        public int SchoolTypeId { get; set; }
+
+        public string Name { get; set; }
+    }
+
+}
