--- DatabaseStructure/SchoolType.cs
@@ -1,11 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations.Schema;
-
-namespace DatabaseStructure
-{
-    public class SchoolType
-    {
-        public int SchoolTypeId { get; set; }
-
-        public string Name { get; set; }
-    }
-}
