--- DatabaseStructure/Student.cs
@@ -1,28 +1,0 @@
-﻿using System;
-using System.Collections.ObjectModel;
-using System.ComponentModel.DataAnnotations.Schema;
-
-namespace DatabaseStructure
-{
-    public class Student
-    {
-        public int StudentId { get; set; }
-
-        public string FirstName { get; set; }
-
-        public string LastName { get; set; }
-
-        public string Email { get; set; }
-
-        public string Phone { get; set; }
-
-        [DatabaseGenerated(DatabaseGeneratedOption.Computed)]
-        public DateTime RegistrationDate { get; set; }
-
-        public bool IsAccountActive { get; set; }
-
-        public virtual School StudentSchool { get; set; }
-
-        public virtual Collection<Course> StudentCourses { get; set; }
-    }
-}
